refactor SearchRenters
